"comparison of integers of different size" fixes
  ?cxx/AssistantCns.cxx:?produceAssistantCns `%s/int x = 0; /size_t x = 0; /`
  ?cxx/VirusAnalysis.cxx:?produceVirusFixCns `%s/int x = 0; /size_t x = 0; /`
  ?cxx/ClassResultList.hxx:?listProduceSignature `%s/size_t smallest;/ptrdiff_t smallest;/`

?posts/VirusAnalysis.md /* include this */

+++ cxx/AssistantCns.cxx
@@ -63,7 +63,7 @@ void produceAssistantCns(const ResultList &questionsOrNull, const ResultList &re
 	cns.setNeuronsPerLayer(26666);
 	assert(questionsOrNull.bytecodes.size() == questionsOrNull.bytecodes.size());
 	inputsToOutputs.reserve(questionsOrNull.bytecodes.size());
-	for(int x = 0; questionsOrNull.bytecodes.size() > x; ++x) {
+	for(size_t x = 0; questionsOrNull.bytecodes.size() > x; ++x) {
 		inputsToOutputs.push_back({questionsOrNull.bytecodes[x], responsesOrNull.bytecodes[x]});
 	}
 	cns.setupSynapses(inputsToOutputs);
